Big refactoring of the packages.
This is mainly to avoid initializing frameworks we don't use.

As a side effect/bonus, I initialize the debug listeners on non-WSGI
processes.
